+++ src/main/kotlin/space/kscience/gradle/KScienceExtension.kt
@@ -3,28 +3,33 @@ package space.kscience.gradle
 import org.gradle.api.Project
 import org.gradle.api.file.DuplicatesStrategy
 import org.gradle.api.plugins.ApplicationPlugin
+import org.gradle.api.plugins.ExtensionAware
 import org.gradle.api.provider.Property
 import org.gradle.api.tasks.Copy
 import org.gradle.api.tasks.testing.Test
 import org.gradle.jvm.toolchain.JavaLanguageVersion
 import org.gradle.kotlin.dsl.*
 import org.gradle.language.jvm.tasks.ProcessResources
+import org.jetbrains.kotlin.gradle.ExperimentalKotlinGradlePluginApi
+import org.jetbrains.kotlin.gradle.ExperimentalWasmDsl
 import org.jetbrains.kotlin.gradle.dsl.KotlinJsProjectExtension
 import org.jetbrains.kotlin.gradle.dsl.KotlinJvmProjectExtension
 import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
 import org.jetbrains.kotlin.gradle.dsl.KotlinProjectExtension
+import org.jetbrains.kotlin.gradle.plugin.KotlinCompilation
 import org.jetbrains.kotlin.gradle.plugin.KotlinDependencyHandler
 import org.jetbrains.kotlin.gradle.plugin.mpp.KotlinNativeTarget
-import org.jetbrains.kotlin.gradle.targets.js.dsl.ExperimentalWasmDsl
 import org.jetbrains.kotlin.gradle.targets.js.dsl.KotlinJsBrowserDsl
 import org.jetbrains.kotlin.gradle.targets.js.dsl.KotlinJsTargetDsl
-import org.jetbrains.kotlin.gradle.targets.js.dsl.KotlinWasmTargetDsl
+import org.jetbrains.kotlin.gradle.targets.js.dsl.KotlinWasmJsTargetDsl
 import org.jetbrains.kotlin.gradle.targets.jvm.KotlinJvmTarget
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 import org.jetbrains.kotlinx.jupyter.api.plugin.tasks.JupyterApiResourcesTask
 import space.kscience.gradle.internal.defaultKotlinJvmArgs
 import space.kscience.gradle.internal.fromJsDependencies
+import space.kscience.gradle.internal.requestPropertyOrNull
 import space.kscience.gradle.internal.useCommonDependency
+import javax.inject.Inject
 
 public enum class DependencyConfiguration {
     API,
@@ -47,9 +52,9 @@ public val Project.isInDevelopment: Boolean
             || version.toString().endsWith("SNAPSHOT")
 
 
-private const val defaultJdkVersion = 11
+private const val defaultJdkVersion = 17
 
-public open class KScienceExtension(public val project: Project) {
+public abstract class KScienceExtension @Inject constructor(public val project: Project): ExtensionAware {
 
     public val jdkVersionProperty: Property<Int> = project.objects.property<Int>().apply {
         set(defaultJdkVersion)
@@ -105,7 +110,7 @@ public open class KScienceExtension(public val project: Project) {
             configure<KotlinMultiplatformExtension> {
                 sourceSets.findByName("commonMain")?.apply {
                     dependencies {
-                        api(platform("io.ktor:ktor-bom:$version"))
+                        api(project.dependencies.platform("io.ktor:ktor-bom:$version"))
                     }
                 }
             }
@@ -114,7 +119,7 @@ public open class KScienceExtension(public val project: Project) {
             configure<KotlinJvmProjectExtension> {
                 sourceSets.findByName("main")?.apply {
                     dependencies {
-                        api(platform("io.ktor:ktor-bom:$version"))
+                        api(project.dependencies.platform("io.ktor:ktor-bom:$version"))
                     }
                 }
             }
@@ -123,7 +128,7 @@ public open class KScienceExtension(public val project: Project) {
             configure<KotlinJsProjectExtension> {
                 sourceSets.findByName("main")?.apply {
                     dependencies {
-                        api(platform("io.ktor:ktor-bom:$version"))
+                        api(project.dependencies.platform("io.ktor:ktor-bom:$version"))
                     }
                 }
             }
@@ -153,14 +158,6 @@ public open class KScienceExtension(public val project: Project) {
             }
         }
 
-        project.pluginManager.withPlugin("org.jetbrains.kotlin.js") {
-            project.configure<KotlinJsProjectExtension> {
-                sourceSets.getByName(sourceSet ?: "main") {
-                    dependencies(dependencyBlock)
-                }
-            }
-        }
-
         project.pluginManager.withPlugin("org.jetbrains.kotlin.multiplatform") {
             project.configure<KotlinMultiplatformExtension> {
                 sourceSets.getByName(sourceSet ?: "commonMain") {
@@ -179,13 +176,6 @@ public open class KScienceExtension(public val project: Project) {
             }
         }
 
-        project.pluginManager.withPlugin("org.jetbrains.kotlin.js") {
-            project.configure<KotlinJsProjectExtension> {
-                sourceSets.getByName(sourceSet ?: "test") {
-                    dependencies(dependencyBlock)
-                }
-            }
-        }
 
         project.pluginManager.withPlugin("org.jetbrains.kotlin.multiplatform") {
             project.configure<KotlinMultiplatformExtension> {
@@ -207,6 +197,7 @@ public open class KScienceExtension(public val project: Project) {
     /**
      * Mark this module as an application module. JVM application should be enabled separately
      */
+    @Deprecated("Use platform-specific applications")
     public fun application() {
         project.extensions.findByType<KotlinProjectExtension>()?.apply {
             explicitApi = null
@@ -231,7 +222,7 @@ public open class KScienceExtension(public val project: Project) {
                 binaries.executable()
             }
 
-            targets.withType<KotlinWasmTargetDsl> {
+            targets.withType<KotlinWasmJsTargetDsl> {
                 binaries.executable()
             }
         }
@@ -242,8 +233,8 @@ public open class KScienceExtension(public val project: Project) {
      */
     public fun useContextReceivers() {
         project.tasks.withType<KotlinCompile> {
-            kotlinOptions {
-                freeCompilerArgs = freeCompilerArgs + "-Xcontext-receivers"
+            compilerOptions{
+              freeCompilerArgs.add("-Xcontext-receivers")
             }
         }
     }
@@ -261,8 +252,8 @@ public open class KScienceExtension(public val project: Project) {
     public val jsTest: DefaultSourceSet get() = DefaultSourceSet("jsTest")
     public val nativeMain: DefaultSourceSet get() = DefaultSourceSet("nativeMain")
     public val nativeTest: DefaultSourceSet get() = DefaultSourceSet("nativeTest")
-    public val wasmMain: DefaultSourceSet get() = DefaultSourceSet("wasmMain")
-    public val wasmTest: DefaultSourceSet get() = DefaultSourceSet("wasmTest")
+    public val wasmJsMain: DefaultSourceSet get() = DefaultSourceSet("wasmJsMain")
+    public val wasmJsTest: DefaultSourceSet get() = DefaultSourceSet("wasmJsTest")
 
 }
 
@@ -285,21 +276,37 @@ public data class KScienceNativeTarget(
         public val linuxX64: KScienceNativeTarget = KScienceNativeTarget(KotlinNativePreset.linuxX64)
         public val mingwX64: KScienceNativeTarget = KScienceNativeTarget(KotlinNativePreset.mingwX64)
         public val macosX64: KScienceNativeTarget = KScienceNativeTarget(KotlinNativePreset.macosX64)
-        public val macosArm64: KScienceNativeTarget = KScienceNativeTarget(KotlinNativePreset.macosX64)
+        public val macosArm64: KScienceNativeTarget = KScienceNativeTarget(KotlinNativePreset.macosArm64)
         public val iosX64: KScienceNativeTarget = KScienceNativeTarget(KotlinNativePreset.iosX64)
         public val iosArm64: KScienceNativeTarget = KScienceNativeTarget(KotlinNativePreset.iosArm64)
         public val iosSimulatorArm64: KScienceNativeTarget = KScienceNativeTarget(KotlinNativePreset.iosSimulatorArm64)
     }
 }
 
-public class KScienceNativeConfiguration {
+public class KScienceNativeConfiguration(private val project: Project) {
 
 
     internal companion object {
-        private fun defaultNativeTargets(): Map<KotlinNativePreset, KScienceNativeTarget> {
-
+        private fun defaultNativeTargets(project: Project): Map<KotlinNativePreset, KScienceNativeTarget> {
             val hostOs = System.getProperty("os.name")
+
+            val targets = project.requestPropertyOrNull("publishing.targets")
+
             return when {
+                targets == "all" -> listOf(
+                    KScienceNativeTarget.linuxX64,
+                    KScienceNativeTarget.mingwX64,
+                    KScienceNativeTarget.macosX64,
+                    KScienceNativeTarget.macosArm64,
+                    KScienceNativeTarget.iosX64,
+                    KScienceNativeTarget.iosArm64,
+                    KScienceNativeTarget.iosSimulatorArm64,
+                )
+
+                targets != null -> {
+                    targets.split(",").map { KScienceNativeTarget(KotlinNativePreset.valueOf(it)) }
+                }
+
                 hostOs.startsWith("Windows") -> listOf(
                     KScienceNativeTarget.linuxX64,
                     KScienceNativeTarget.mingwX64
@@ -314,6 +321,7 @@ public class KScienceNativeConfiguration {
                 )
 
                 hostOs == "Linux" -> listOf(KScienceNativeTarget.linuxX64)
+
                 else -> {
                     emptyList()
                 }
@@ -322,7 +330,7 @@ public class KScienceNativeConfiguration {
     }
 
 
-    internal var targets: Map<KotlinNativePreset, KScienceNativeTarget> = defaultNativeTargets()
+    internal var targets: Map<KotlinNativePreset, KScienceNativeTarget> = defaultNativeTargets(project)
 
 
     /**
@@ -346,7 +354,8 @@ public class KScienceNativeConfiguration {
     ): Unit = target(KScienceNativeTarget(preset, targetName, targetConfiguration))
 }
 
-public open class KScienceMppExtension(project: Project) : KScienceExtension(project) {
+public abstract class KScienceMppExtension @Inject constructor(project: Project) : KScienceExtension(project) {
+
     /**
      * Enable jvm target
      */
@@ -354,10 +363,9 @@ public open class KScienceMppExtension(project: Project) : KScienceExtension(pro
         project.pluginManager.withPlugin("org.jetbrains.kotlin.multiplatform") {
             project.configure<KotlinMultiplatformExtension> {
                 jvm {
-                    compilations.all {
-                        kotlinOptions {
-                            freeCompilerArgs = freeCompilerArgs + defaultKotlinJvmArgs
-                        }
+                    @OptIn(ExperimentalKotlinGradlePluginApi::class)
+                    compilerOptions{
+                        freeCompilerArgs.addAll(defaultKotlinJvmArgs)
                     }
                     block()
                 }
@@ -387,6 +395,7 @@ public open class KScienceMppExtension(project: Project) : KScienceExtension(pro
             project.configure<KotlinMultiplatformExtension> {
                 js(IR) {
                     browser()
+                    useEsModules()
                     block()
                 }
                 sourceSets {
@@ -403,13 +412,30 @@ public open class KScienceMppExtension(project: Project) : KScienceExtension(pro
         }
     }
 
+    /**
+     * Add Wasm/Js target
+     */
     @OptIn(ExperimentalWasmDsl::class)
-    public fun wasm(block: KotlinWasmTargetDsl.() -> Unit = {}) {
+    public fun wasm(block: KotlinWasmJsTargetDsl.() -> Unit = {}) {
+
         project.pluginManager.withPlugin("org.jetbrains.kotlin.multiplatform") {
             project.configure<KotlinMultiplatformExtension> {
-                wasm {
-                    browser()
+                wasmJs {
+                    browser {
+                        testTask {
+                            useKarma {
+                                useChromeHeadless()
+                            }
+                        }
+                    }
                     block()
+                }
+                sourceSets {
+                    getByName("wasmJsTest") {
+                        dependencies {
+                            implementation(kotlin("test-wasm-js"))
+                        }
+                    }
                 }
             }
         }
@@ -425,26 +451,30 @@ public open class KScienceMppExtension(project: Project) : KScienceExtension(pro
      */
     public fun fullStack(
         bundleName: String = "js/bundle.js",
+        development: Boolean = false,
         jvmConfig: KotlinJvmTarget.() -> Unit = {},
         jsConfig: KotlinJsTargetDsl.() -> Unit = {},
         browserConfig: KotlinJsBrowserDsl.() -> Unit = {},
     ) {
         js {
-            binaries.executable()
             browser {
-                webpackTask {
+                commonWebpackConfig{
                     outputFileName = bundleName
                 }
                 browserConfig()
             }
+            useEsModules()
             jsConfig()
+            binaries.executable()
         }
         jvm {
             val processResourcesTaskName =
-                compilations[org.jetbrains.kotlin.gradle.plugin.KotlinCompilation.MAIN_COMPILATION_NAME]
-                    .processResourcesTaskName
+                compilations[KotlinCompilation.MAIN_COMPILATION_NAME].processResourcesTaskName
 
-            val jsBrowserDistribution = project.tasks.getByName("jsBrowserDistribution")
+
+            val jsBrowserDistribution = project.tasks.getByName(
+                if (development) "jsBrowserDevelopmentExecutableDistribution" else "jsBrowserDistribution"
+            )
 
             project.tasks.getByName<ProcessResources>(processResourcesTaskName) {
                 duplicatesStrategy = DuplicatesStrategy.WARN
@@ -459,90 +489,59 @@ public open class KScienceMppExtension(project: Project) : KScienceExtension(pro
      * Enable all supported native targets
      */
     public fun native(block: KScienceNativeConfiguration.() -> Unit = {}): Unit = with(project) {
-        val nativeConfiguration = KScienceNativeConfiguration().apply(block)
+        val nativeConfiguration = KScienceNativeConfiguration(this).apply(block)
         pluginManager.withPlugin("org.jetbrains.kotlin.multiplatform") {
             configure<KotlinMultiplatformExtension> {
-                val nativeTargets: List<KotlinNativeTarget> =
-                    nativeConfiguration.targets.values.map { nativeTarget ->
-                        when (nativeTarget.preset) {
-                            KotlinNativePreset.linuxX64 -> linuxX64(
-                                nativeTarget.targetName,
-                                nativeTarget.targetConfiguration
-                            )
+                nativeConfiguration.targets.values.forEach { nativeTarget ->
+                    when (nativeTarget.preset) {
+                        KotlinNativePreset.linuxX64 -> linuxX64(
+                            nativeTarget.targetName,
+                            nativeTarget.targetConfiguration
+                        )
 
-                            KotlinNativePreset.mingwX64 -> mingwX64(
-                                nativeTarget.targetName,
-                                nativeTarget.targetConfiguration
-                            )
+                        KotlinNativePreset.mingwX64 -> mingwX64(
+                            nativeTarget.targetName,
+                            nativeTarget.targetConfiguration
+                        )
 
-                            KotlinNativePreset.macosX64 -> macosX64(
-                                nativeTarget.targetName,
-                                nativeTarget.targetConfiguration
-                            )
+                        KotlinNativePreset.macosX64 -> macosX64(
+                            nativeTarget.targetName,
+                            nativeTarget.targetConfiguration
+                        )
 
-                            KotlinNativePreset.macosArm64 -> macosArm64(
-                                nativeTarget.targetName,
-                                nativeTarget.targetConfiguration
-                            )
+                        KotlinNativePreset.macosArm64 -> macosArm64(
+                            nativeTarget.targetName,
+                            nativeTarget.targetConfiguration
+                        )
 
-                            KotlinNativePreset.iosX64 -> iosX64(
-                                nativeTarget.targetName,
-                                nativeTarget.targetConfiguration
-                            )
+                        KotlinNativePreset.iosX64 -> iosX64(
+                            nativeTarget.targetName,
+                            nativeTarget.targetConfiguration
+                        )
 
-                            KotlinNativePreset.iosArm64 -> iosArm64(
-                                nativeTarget.targetName,
-                                nativeTarget.targetConfiguration
-                            )
+                        KotlinNativePreset.iosArm64 -> iosArm64(
+                            nativeTarget.targetName,
+                            nativeTarget.targetConfiguration
+                        )
 
-                            KotlinNativePreset.iosSimulatorArm64 -> iosSimulatorArm64(
-                                nativeTarget.targetName,
-                                nativeTarget.targetConfiguration
-                            )
-
-//                                else -> {
-//                                    logger.error("Native preset ${nativeTarget.preset} not recognised.")
-//                                    null
-//                                }
-                        }
-                    }
-
-                sourceSets {
-                    val commonMain by getting
-                    val commonTest by getting
-
-                    val nativeMain by creating {
-                        dependsOn(commonMain)
-                    }
-
-                    val nativeTest by creating {
-                        //should NOT depend on nativeMain because automatic dependency by plugin
-                        dependsOn(commonTest)
-                    }
-
-                    configure(nativeTargets) {
-                        compilations["main"]?.apply {
-                            configure(kotlinSourceSets) {
-                                dependsOn(nativeMain)
-                            }
-                        }
-
-                        compilations["test"]?.apply {
-                            configure(kotlinSourceSets) {
-                                dependsOn(nativeTest)
-                            }
-                        }
+                        KotlinNativePreset.iosSimulatorArm64 -> iosSimulatorArm64(
+                            nativeTarget.targetName,
+                            nativeTarget.targetConfiguration
+                        )
                     }
                 }
+                applyDefaultHierarchyTemplate()
             }
         }
     }
 }
 
 
-internal inline fun <reified T : KScienceExtension> Project.registerKScienceExtension(constructor: (Project) -> T): T {
-    extensions.findByType<T>()?.let { return it }
-    return constructor(this).also {
-        extensions.add("kscience", it)
-    }
+internal inline fun <reified T : KScienceExtension> Project.registerKScienceExtension():T {
+//    extensions.findByType<T>()?.let { return it }
+//    return constructor(this).also {
+//        extensions.add("kscience", it)
+//    }
+
+    return extensions.create("kscience", T::class.java)
 }