+++ src/main/kotlin/space/kscience/gradle/KScienceJVMPlugin.kt
@@ -10,23 +10,27 @@ import org.gradle.kotlin.dsl.get
 import org.gradle.kotlin.dsl.withType
 import org.jetbrains.dokka.gradle.DokkaPlugin
 import org.jetbrains.kotlin.gradle.dsl.KotlinJvmProjectExtension
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 import space.kscience.gradle.internal.applySettings
 import space.kscience.gradle.internal.defaultKotlinJvmArgs
 
 public open class KScienceJVMPlugin : Plugin<Project> {
     override fun apply(project: Project): Unit = project.run {
+
+        logger.warn("KSCience JVM plugin is deprecated. Use MPP.")
         if (!plugins.hasPlugin("org.jetbrains.kotlin.jvm")) {
             plugins.apply("org.jetbrains.kotlin.jvm")
         } else {
             logger.info("Kotlin JVM plugin is already present")
         }
-        val extension = registerKScienceExtension(::KScienceExtension)
+        val extension = registerKScienceExtension<KScienceExtension>()
 
         //logger.info("Applying KScience configuration for JVM project")
         configure<KotlinJvmProjectExtension> {
             sourceSets.all {
                 languageSettings.applySettings()
+                compilerOptions{
+                    freeCompilerArgs.addAll(defaultKotlinJvmArgs)
+                }
             }
 
             sourceSets["test"].apply {
@@ -39,12 +43,6 @@ public open class KScienceJVMPlugin : Plugin<Project> {
             if (explicitApi == null) explicitApiWarning()
             jvmToolchain {
                 languageVersion.set(extension.jdkVersionProperty.map { JavaLanguageVersion.of(it) })
-            }
-        }
-
-        tasks.withType<KotlinCompile> {
-            kotlinOptions {
-                freeCompilerArgs = freeCompilerArgs + defaultKotlinJvmArgs
             }
         }