+++ src/main/kotlin/space/kscience/gradle/KScienceProjectPlugin.kt
@@ -12,7 +12,12 @@ import org.jetbrains.changelog.ChangelogPlugin
 import org.jetbrains.changelog.ChangelogPluginExtension
 import org.jetbrains.dokka.gradle.AbstractDokkaTask
 import org.jetbrains.dokka.gradle.DokkaPlugin
-import space.kscience.gradle.internal.*
+import org.jetbrains.kotlin.gradle.targets.js.yarn.YarnLockMismatchReport
+import org.jetbrains.kotlin.gradle.targets.js.yarn.YarnPlugin
+import org.jetbrains.kotlin.gradle.targets.js.yarn.YarnRootExtension
+import space.kscience.gradle.internal.addPublishing
+import space.kscience.gradle.internal.addSonatypePublishing
+import space.kscience.gradle.internal.setupPublication
 
 /**
  * Simplifies adding repositories for Maven publishing, responds for releasing tasks for projects.
@@ -51,46 +56,24 @@ public class KSciencePublishingExtension(public val project: Project) {
     }
 
     /**
-     * Adds GitHub as VCS and adds GitHub Packages Maven repository to publishing.
-     *
-     * @param githubProject the GitHub project.
-     * @param githubOrg the GitHub user or organization.
-     * @param deploy publish packages in the `deploy` task to the GitHub repository.
+     * Add a repository with [repositoryName]. Uses "publishing.$repositoryName.user" and "publishing.$repositoryName.token"
+     * properties pattern to store user and token
      */
-    public fun github(
-        githubOrg: String,
-        githubProject: String,
-        deploy: Boolean = project.requestPropertyOrNull("publishing.github") == "true",
+    public fun repository(
+        repositoryName: String,
+        url: String,
     ) {
-        if (deploy) {
-            try {
-                project.addGithubPublishing(githubOrg, githubProject)
-            } catch (t: Throwable) {
-                project.logger.error("Failed to set up github publication", t)
-            }
-        }
-    }
-
-    /**
-     * Adds Space Packages Maven repository to publishing.
-     *
-     * @param spaceRepo the repository URL.
-     * @param deploy publish packages in the `deploy` task to the Space repository.
-     */
-    public fun space(
-        spaceRepo: String,
-    ) {
-        project.addSpacePublishing(spaceRepo)
+        require(isVcsInitialized) { "The project vcs is not set up use 'pom' method to do so" }
+        project.addPublishing(repositoryName, url)
     }
 
     /**
      * Adds Sonatype Maven repository to publishing.
-     *
-     * @param addToRelease  publish packages in the `release` task to the Sonatype repository.
+
      */
-    public fun sonatype() {
+    public fun sonatype(sonatypeRoot: String = "https://s01.oss.sonatype.org") {
         require(isVcsInitialized) { "The project vcs is not set up use 'pom' method to do so" }
-        project.addSonatypePublishing()
+        project.addSonatypePublishing(sonatypeRoot)
     }
 }
 
@@ -104,11 +87,11 @@ public open class KScienceProjectPlugin : Plugin<Project> {
         apply<DokkaPlugin>()
         apply<BinaryCompatibilityValidatorPlugin>()
 
-        allprojects{
-            repositories{
+        allprojects {
+            repositories {
                 mavenCentral()
                 maven("https://repo.kotlin.link")
-                maven("https://maven.pkg.jetbrains.space/spc/p/sci/dev")
+                google()
             }
 
             // Workaround for https://github.com/gradle/gradle/issues/15568
@@ -139,10 +122,11 @@ public open class KScienceProjectPlugin : Plugin<Project> {
             val readmeExtension = KScienceReadmeExtension(this)
             extensions.add("readme", readmeExtension)
 
-            @Suppress("UNUSED_VARIABLE")
-            val generateReadme by tasks.creating {
+            tasks.create("generateReadme") {
                 group = "documentation"
                 description = "Generate a README file if stub is present"
+
+                inputs.property("extension", readmeExtension)
 
                 if (readmeExtension.readmeTemplate.exists()) {
                     inputs.file(readmeExtension.readmeTemplate)
@@ -179,6 +163,8 @@ public open class KScienceProjectPlugin : Plugin<Project> {
                 }
             }
 
+            inputs.property("extension", rootReadmeExtension)
+
             if (rootReadmeExtension.readmeTemplate.exists()) {
                 inputs.file(rootReadmeExtension.readmeTemplate)
             }
@@ -206,7 +192,7 @@ public open class KScienceProjectPlugin : Plugin<Project> {
                             subproject.extensions.findByType<KScienceReadmeExtension>()?.let { ext ->
                                 val path = subproject.path.replaceFirst(":", "").replace(":", "/")
                                 appendLine("\n### [$path]($path)")
-                                appendLine("> ${ext.description}")
+                                ext.description?.let { appendLine("> ${ext.description}") }
                                 appendLine(">\n> **Maturity**: ${ext.maturity}")
                                 val featureString = ext.featuresString(itemPrefix = "> - ", pathPrefix = "$path/")
                                 if (featureString.isNotBlank()) {
@@ -233,12 +219,12 @@ public open class KScienceProjectPlugin : Plugin<Project> {
 
         tasks.create("version") {
             group = "publishing"
-            val versionFile = project.buildDir.resolve("project-version.txt")
-            outputs.file(versionFile)
+            val versionFileProvider = project.layout.buildDirectory.file("project-version.txt")
+            outputs.file(versionFileProvider)
             doLast {
+                val versionFile = versionFileProvider.get().asFile
                 versionFile.createNewFile()
                 versionFile.writeText(project.version.toString())
-                println(project.version)
             }
         }
 
@@ -250,9 +236,11 @@ public open class KScienceProjectPlugin : Plugin<Project> {
             }
         }
 
-        plugins.withType(org.jetbrains.kotlin.gradle.targets.js.yarn.YarnPlugin::class.java) {
-            rootProject.the<org.jetbrains.kotlin.gradle.targets.js.yarn.YarnRootExtension>().lockFileDirectory =
-                rootDir.resolve("gradle")
+        plugins.withType<YarnPlugin>() {
+            rootProject.configure<YarnRootExtension> {
+                lockFileDirectory = rootDir.resolve("gradle")
+                yarnLockMismatchReport = YarnLockMismatchReport.WARNING
+            }
         }
     }